Implement Msg DateTime and Sequence only in JetStream
