+++ .editorconfig
@@ -13,6 +13,7 @@ generated_code = true
 
 [*.cs]
 indent_size = 4
+max_line_length = 300
 
 # changes from VS2019 defaults
 csharp_style_namespace_declarations = file_scoped
